Fix 4.11.0 launch issue, and some clean up.
Only /usr/lib/webcord is necessary for the app to function, as with pretty much any Electron app. Therefore it should be safe to delete everything else to save a bit of space. This also fixes the launch issue on 4.11.0 as the Electron version used by that build seems to no longer search for the lib directory.